Correct minor bug to properly support OUTPUT ducktyping
XML::Writer *Almost* supports "Duck Typing" for the OUTPUT parameter,
and in fact, includes it's own inline package with a 'print' method for
the case where someone passes in a scalar reference.

This patch adds documentation, test cases, and a minor code patch to
support passing in any arbitrary object that has a print() method.
Taking an cue from Plack's recommendations, I'm purposely not checking
if the object has a print() method defined.